+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - cli: Improve dashboard command output
+- cli: Increase max todos
 
 ## [v0.6.0] - 2025-08-27